requirements.txt
streamlit
base64
pathlib
pillow